GM-PHD Filter

Gaussian Mixture Probability Hypothesis Density Filter

A Gaussian Mixture Probability Hypothesis Density (GM-PHD) Filter handles multi-object tracking in a low signal-to-noise environment by recursively propagating an unnormalized Gaussian Mixture Model, which when integrated over a volume corresponds with the expected number of objects in that volume.

GaussianFilters.GaussianMixture — Type.

GaussianMixture(N::Int64, w::Vector{Float64}, μ::Vector{Vector},
    Σ::Vector{Matrix})
GaussianMixture(w::Vector{Float64}, μ::Vector{Vector}, Σ::Vector{Matrix})

Arguments:
N: Number of models
w: Weights
μ: Means of the model
Σ: Covariances of the model

source

Building a GM-PHD Filter

Building a GM-PHD Filter requires defining a birth intensity model of type GaussianMixture for where objects can be globally born, a spawn intensity model of type Spawn for how new objects can spawn off old objects, a Vector of possible linear dynamics models of type Dynamics, a linear measurement model of type Measurement, a survival probability, a detection probability, and a clutter modeling function.

GaussianFilters.Spawn — Type.

Spawn(β::GaussianMixture, dyn::Vector{Dynamics})

Arguments:
β: Gaussian Mixture model determining the
spawning intesity of the target
dyn: Dynamics

source

GaussianFilters.Dynamics — Type.

Dynamics(A::Matrix, Q::Matrix, d::Vector)
Dynamics(A::Matrix, Q::Matrix)

Construct linear dynamics model with; transition matrix A, process noise matrix Q and offset vector d

source

GaussianFilters.Measurement — Type.

Measurement(C::Matrix, R::Matrix)

Construct measurement model with observation matrix C and sensor noise matrix R

source

Once all of these are defined, a GM-PHD filter can be constructed with PHDFilter.

GaussianFilters.PHDFilter — Type.

PHDFilter(γ::GaussianMixture, spawn::Spawn, dyn::Vector{Dynamics},
    meas::Measurement, Ps::Float64, Pd::Float64, κ::Function)
PHDFilter(γ::GaussianMixture, spawn::Spawn, dyn::Dynamics,
    meas::Measurement, Ps::Float64, Pd::Float64, κ::Function)

Sets up a PHD Filter

Arguments:
γ: Birth intensity
spawn: Spawning intensity
dyn: Vector of possible Dynamics models
meas: Measurements
Ps: Survival probability
Pd: Detection probability

source

Running a GM-PHD Filter

Currently, the GM-PHD Filter must be run step-by-step. Prediction and measurement steps are combined in step, while resulting low-probability mixture components are pruned with prune. These functions are combined in step_prune.

Base.step — Function.

Function to combine stepping through PHD filter and pruning resultant distribution

Arguments:

x::GaussianMixture Prior state distribution. [Gaussian Mixture]

Z::Matrix Matrix of measurements. [Measurements]

PHD::PHDFilter PHD filter to step through. [PHD Filter]

T::Real Threshholding value for pruning. [probability]

Returns:

xp::GaussianMixture Describe first return value. [Gaussian Mixture]

References:

Vo, B. N., & Ma, W. K. (2006). The Gaussian mixture probability hypothesis

density filter. IEEE Transactions on signal processing, 54(11), 4091-4104.

source

GaussianFilters.prune — Function.

prune(x::GaussianMixture, T::Real, U::Real, J_max::Integer)

Prune the posterior Gaussian-Mixture

Arguments:

x::GaussianMixture Posterior state distribution. [Gaussian Mixture]

T::Real Truncation threshold. Drop distributions with weight less than T

U::Real Merging threshold. Merge if (μ1-μ2)^T * Σ^-1 * (μ1-μ2) < U

J_max::Integer Maximum number of features

source

GaussianFilters.step_prune — Function.

Function to combine stepping through PHD filter and pruning resultant distribution

Arguments:

x::GaussianMixture Prior state distribution. [Gaussian Mixture]

Z::Matrix Matrix of measurements. [Measurements]

PHD::PHDFilter PHD filter to step through. [PHD Filter]

T::Real Threshholding value for pruning. [probability]

Returns:

xp::GaussianMixture Describe first return value. [Gaussian Mixture]

References:

Vo, B. N., & Ma, W. K. (2006). The Gaussian mixture probability hypothesis

density filter. IEEE Transactions on signal processing, 54(11), 4091-4104.

source

Target locations can be extracted from a GaussianMixture state using multiple_target_state_extraction.

GaussianFilters.multiple_target_state_extraction — Function.

multiple_target_state_extraction(x, threshold)

Extracts targets whose weights (x.w) are above threshold.

Arguments: x: Set of Gaussian Mixtures threshold: Threshold on weights. Above this threshold, state estimate is extracted

Returns: X: Multi-Target State Estimate

source

Examples

Full implementation examples can be found in the notebooks folder of the repo:

GM-PHD Object Surveillance Example

GM-PHD Aircraft Carrier Example