Kalman-class Filters

Kalman-Class Filters

The Kalman, Extended Kalman, and Unscented Kalman filters are used to estimate state using unimodal multivariate Gaussian distributions. A state estimate in this framework is defined as a GaussianBelief consisting of a mean and covariance.

GaussianFilters.GaussianBelief — Type.

GaussianBelief(μ::AbstractVector,Σ::Symmetric)
GaussianBelief(μ::AbstractVector,Σ::AbstractMatrix)

Construct a gaussian belief, consisting of mean vector μ and symmetric covariance matrix Σ

source

Building a Filter

In general, Kalman-class filters can be built with either linear or non-linear dynamics and measurement models. Linear models should be defined with appropriately matrices. Non-linear models should be defined using an appropriate function of two variables, state and action. Both models should be defined with symmetric noise covariance matrices.

NOTE: There is no need to define Jacobians for non-linear models, since this package uses automatic forward differentiation to compute Jacobians in real time. Just make sure the models are forward differentiable in all possible belief locations.

GaussianFilters.LinearDynamicsModel — Type.

LinearDynamicsModel(A::AbstractMatrix,B::AbstractMatrix,W::Symmetric)
LinearDynamicsModel(A::AbstractMatrix,B::AbstractMatrix,W::AbstractMatrix)

Construct linear dynamics model with; transition matrix A, control matrix B, and symmetric zero-mean process noise with symmetric covariance matrix W

source

GaussianFilters.LinearObservationModel — Type.

LinearObservationModel(C::AbstractMatrix,D::AbstractMatrix,V::Symmetric)
LinearObservationModel(C::AbstractMatrix,D::AbstractMatrix,V::AbstractMatrix)
LinearObservationModel(C::AbstractMatrix,V::Symmetric)
LinearObservationModel(C::AbstractMatrix,V::AbstractMatrix)

Construct linear observation dynamics model with; transition matrix C, control matrix B, and symmetric zero-mean measurement noise with symmetric covariance matrix V

source

GaussianFilters.NonlinearDynamicsModel — Type.

NonlinearDynamicsModel(f::Function,W::Symmetric)
NonlinearDynamicsModel(f::Function,W::AbstractMatrix)

Construct nonlinear dynamics model with transition function f and symmetric zero-mean process noise with symmetric covariance matrix W

source

GaussianFilters.NonlinearObservationModel — Type.

NonlinearObservationModel(h::Function,V::Symmetric)
NonlinearObservationModel(h::Function,V::AbstractMatrix)

Construct nonlinear observation dynamics model with measurement function h and symmetric zero-mean measurement noise with symmetric covariance matrix V

source

Use the filter constructors with the appropriately typed models to build a filter. It is recommended to always construct a Kalman filter type when both dynamics and observation models are linear.

GaussianFilters.KalmanFilter — Type.

KalmanFilter(d::LinearDynamicsModel,o::LinearObservationModel)

Construct Kalman filter with LinearDynamicsModel d and LinearObservationModel o.

source

GaussianFilters.ExtendedKalmanFilter — Type.

ExtendedKalmanFilter(d::DynamicsModel,o::ObservationModel)
KalmanFilter(d::DynamicsModel,o::ObservationModel,λ::Number,
    α::Float,β::Float)

Construct Extended Kalman filter with DynamicsModel d and ObservationModel o.

source

GaussianFilters.UnscentedKalmanFilter — Type.

UnscentedKalmanFilter(d::DynamicsModel,o::ObservationModel,λ::Number,
    α::Float,β::Float)
UnscentedKalmanFilter(d::DynamicsModel,o::ObservationModel,λ::Number)
UnscentedKalmanFilter(d::DynamicsModel,o::ObservationModel)

Construct Unscented Kalman filter with DynamicsModel d, ObservationModel o, and UKF parameters λ, α, and β. Default constructor uses α/β formulation from Probabilistic Robotics, second constructor reduces complexity, third constructor defaults λ to 2, as is commonly done.

source

Simulating Data

Given an initial belief, an action sequence, and a filter, you can either simulate state and measurement data all at once with simulation or one step at a time with simulate_step

GaussianFilters.simulation — Function.

simulation(b0::GaussianBelief,action_sequence::Vector{AbstractVector}},
    filter::AbstractFilter)

Run a simulation to get positions and measurements. Samples starting point from GaussianBelief b0, the runs action_sequence with additive gaussian noise all specified by AbstractFilter filter to return a simulated state and measurement history.

source

GaussianFilters.simulate_step — Function.

simulate_step(x::AbstractVector, u::AbstractVector, filter::KalmanFilter)

Run a step of simulation starting at state x, taking action u, and using the motion and measurement equations specified by Kalman Filter filter.

source

simulate_step(x::AbstractVector, u::AbstractVector, filter::ExtendedKalmanFilter)

Run a step of simulation starting at state x, taking action u, and using the motion and measurement equations specified by Kalman Filter filter.

source

simulate_step(x::AbstractVector, u::AbstractVector, filter::UnscentedKalmanFilter)

Run a step of simulation starting at state x, taking action u, and using the motion and measurement equations specified by Kalman Filter filter.

source

Running a Filter

You can run a filter on a sequential measurement data using the run_filter function.

GaussianFilters.run_filter — Function.

run_filter(b0::GaussianBelief, action_history::Vector{AbstractVector},
        measurement_history::Vector{AbstractVector}, filter::AbstractFilter)

Given an initial belief b0, matched-size arrays for action and measurement histories and a filter, update the beliefs using the filter, and return a vector of all beliefs.

source

Alternatively, you can make step-wise belief updates using the update function, which consists of a two-step process to a) predict the next state given a known action and b) make measurement-based belief updates with measure.

GaussianFilters.update — Function.

update(b0::GaussianBelief, u::AbstractVector, y::AbstractVector,
    filter::AbstractFilter)

Uses AbstractFilter filter to update gaussian belief b0, given control vector u and measurement vector y.

source

update(x::GaussianMixture, Z::Vector{AbstractVector}, PHD::PHDFilter,
	T::Real, U::Real, J_max::Integer)

Perform an update step using a GMPHD Filter.

Arguments:

x::GaussianMixture Prior state distribution. [Gaussian Mixture]

Z::Matrix Matrix of measurements. [Measurements]

PHD::PHDFilter PHD filter to step through. [PHD Filter]

T::Real Truncation threshold. Drop distributions with weight less than T

U::Real Merging threshold. Merge if (μ1-μ2)^T * Σ^-1 * (μ1-μ2) < U

J_max::Integer Maximum number of features

Returns:

xn::GaussianMixture Describe first return value. [Gaussian Mixture]

References:

Vo, B. N., & Ma, W. K. (2006). The Gaussian mixture probability hypothesis

density filter. IEEE Transactions on signal processing, 54(11), 4091-4104.

source

GaussianFilters.predict — Function.

predict(b0::GaussianBelief, u::AbstractVector, filter::KalmanFilter)

Uses Kalman filter to run prediction step on gaussian belief b0, given control vector u.

source

predict(b0::GaussianBelief, u::AbstractVector, filter::ExtendedKalmanFilter)

Uses Extended Kalman filter to run prediction step on gaussian belief b0, given control vector u.

source

predict(b0::GaussianBelief, u::AbstractVector, filter::UnscentedKalmanFilter)

Uses Unscented Kalman filter to run prediction step on gaussian belief b0, given control vector u.

source

predict(x::GaussianMixture, PHD::PHDFilter)

Make a prediction on next state based on PHD dynamics

Arguments:

x::GaussianMixture Prior state distribution. [Gaussian Mixture]

PHD::PHDFilter PHD filter to step through. [PHD Filter]

Returns:

xp::GaussianMixture Predicted next state distribution. [Gaussian Mixture]

References:

Vo, B. N., & Ma, W. K. (2006). The Gaussian mixture probability hypothesis

density filter. IEEE Transactions on signal processing, 54(11), 4091-4104.

source

GaussianFilters.measure — Function.

measure(bp::GaussianBelief, y::AbstractVector, filter::KalmanFilter;
    u::AbstractVector = [false])

Uses Kalman filter to run measurement update on predicted gaussian belief bp, given measurement vector y. If u is specified and filter.o.D has been declared, then matrix D will be factored into the y predictions

source

measure(bp::GaussianBelief, y::AbstractVector, filter::ExtendedKalmanFilter;
    u::AbstractVector = [false])

Uses Extended Kalman filter to run measurement update on predicted gaussian belief bp, given measurement vector y. If u is specified and filter.o.D has been declared, then matrix D will be factored into the y predictions.

source

measure(bp::GaussianBelief, y::AbstractVector, filter::UnscentedKalmanFilter;
    u::AbstractVector = [false])

Uses Unscented Kalman filter to run measurement update on predicted gaussian belief bp, given measurement vector y. If u is specified and filter.o.D has been declared, then matrix D will be factored into the y predictions.

source

measure(xp::GaussianMixture, Z::Vector{AbstractVector}, PHD::PHDFilter)

Perform a measurement update on a predicted PHD next state.

Arguments:

xp::GaussianMixture Prior state distribution. [Gaussian Mixture]

Z::Vector{AbstractVector} Array of measurements. [Measurements]

PHD::PHDFilter PHD filter to step through. [PHD Filter]

Returns:

xm::GaussianMixture Describe first return value. [Gaussian Mixture]

References:

Vo, B. N., & Ma, W. K. (2006). The Gaussian mixture probability hypothesis

density filter. IEEE Transactions on signal processing, 54(11), 4091-4104.

source

Utilities

The output of running a filter is a GaussianBelief vector, which can be condensed into appropriate tensors with unpack.

GaussianFilters.unpack — Function.

unpack(belief_history::Vector{GaussianBelief};
    dims::Vector{Int}=[])

Given a history of beliefs, return an unpacked (time steps, state dim)-sized array of predicted means and a (time steps, state dim, state dim)-sized array of covariances. One can optionally specify dimensions indices dims to output reduced state information.

source

belief_ellipse can be used to convert a 2-D Gaussian belief into points along a confidence interval ellipse for plotting.

GaussianFilters.belief_ellipse — Function.

belief_ellipse(b::GaussianBelief, P::Float=0.95; δ::Number=5)

Construct and return the x and y points of a 2D gaussian belief, with P being the total probability captured by the ellipse (P ∈ (0,1)), and δ the degree increment between points.

source

Examples

Full implementation examples can be found in the notebooks folder of the repo:

Kalman Filter Example

Extended Kalman Filter Example

Unscented Kalman Filter Example